requirements.txt
# Core orchestration
langchain==0.1.0
langchain-core==0.1.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Authentication
pyjwt>=2.8.0

# Pydantic (config validation)
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM Providers
anthropic==0.25.0
google-generativeai==0.3.0
openai==1.3.0

# Database
supabase==2.3.0
postgrest-py==0.13.0

# Async
aiohttp==3.9.0

# Config & ENV
python-dotenv==1.0.0
pyyaml==6.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.0

# Docker
docker-compose>=1.29.0
